requirements/emu2.txt
torch
transformers==4.31.0
peft==0.6.2
numpy
Pillow
argparse
einops
timm
xformers
safetensors
sentencepiece==0.1.99
decord
diffusers==0.24.0
torchvision
accelerate==0.25.0
gradio==3.40.1
tokenizers==0.13.3
protobuf
accelerate
bitsandbytes
python-Levenshtein